TrafficCapture/dockerSolution/src/main/docker/migrationConsole/console_api/requirements.txt
-e ../lib/console_link

Django==4.2.13
django-extensions==3.2.3
djangorestframework==3.15.1
Werkzeug==3.0.3

# Vulnerable library fixes
sqlparse==0.5.0